Estimate the probability of detection, given a location is occupied, based on the audio survey length and the number of surveys conducted.

Using data from the ABMI ecosystem health acoustic surveys (2015-2020), occupancy models were fit to five common boreal songbirds. These models are used to estimate the detection probability for each species for surveys of different lengths (1-3 minutes). Given that it is present, these results can be used to estimate the probability the species will be detected at least once as a function of the number of surveys conducted.

wt_prob_det(species_code, survey_length, number_of_surveys)

Arguments

species_code
Character; the species code. See below for included species.

survey_length
Double; the survey length in minutes. Valid values are 1, 2, and 3.

number_of_surveys
Double; the number of surveys (to be) conducted.

Value

Double; the estimated probability of species detection based on the number of surveys planned.

Details

Valid values for species_code currently include:

"OVEN" (Ovenbird)

"CCSP" (Clay-colored Sparrow)

"OSFL" (Olive-sided Flycatcher)

"TEWA" (Tennessee Warbler)

"WTSP" (White-throated Sparrow)

Examples

if (FALSE) {
wt_prob_det(species_code = "OVEN", survey_length = 1, number_of_surveys = 4)
}