Set of analysis functions

This function takes your independent detection data and summarises it by location, specified time interval, and species.

Usage

wt_summarise_cam(
  detect_data,
  raw_data,
  time_interval = "day",
  variable = "detections",
  output_format = "wide",
  species_col = species_common_name,
  effort_data = NULL,
  project_col = NULL,
  station_col = NULL,
  start_col = NULL,
  end_col = NULL
)

Arguments

detect_data

Detection data generated from wt_ind_det().

raw_data

The raw camera tag data, which is used to infer the effort (i.e. date ranges of operation) for each camera. Optionally, can supply effort_data directly instead.

time_interval

Character; Can be either "full", "month", "week", or "day".

variable

Character; Can be either "presence", "detections", "counts", or "all" (if you want all three).

output_format

Character; The format of the dataframe returned to you. Can be either "wide" or "long".

species_col

Defaults to common_name. The column referring to species. Use to switch between common and scientific names of species, if you have both.

effort_data

Optionally supply your own effort data.

project_col

The column referring to project in your effort data.

station_col

The column referring to each individual camera station/location in your effort data.

start_col

The column indicating the start date of the camera location

end_col

The column indicating the end date of the camera location

Value

A dataframe summarising your camera data by location, time interval, and species.

Summarise your camera data by location, time interval, and species.

NA

Examples

if (FALSE) {
summary <- wt_summarise_cam(
x, y, time_interval = "day", variable = "detections", output_format = "wide"
)
}