Download Reports

Download ARU, Camera, or Point Count data from a project

Usage

wt_download_report(project_id, sensor_id, reports, weather_cols = TRUE)

Arguments

project_id

Numeric; the project ID number that you would like to download data for. Use wt_get_download_summary() to retrieve these IDs.

sensor_id

Character; Can either be "ARU", "CAM", or "PC".

reports

Character; The report type to be returned. Multiple values are accepted as a concatenated string.

weather_cols

Logical; Do you want to include weather information for your stations? Defaults to TRUE.

Value

If multiple report types are requested, a list object is returned; if only one, a dataframe.

Details

Valid values for argument report when sensor_id = "CAM" currently are:

main

project

location

image_report

image_set

tag

megadetector

megaclassifier

definitions

Valid values for argument report when sensor_id = "ARU" currently are:

main

project

location

recording

tag

birdnet

definitions

Valid values for argument report when sensor_id = "PC" currently are:

main

project

location

point count

definitions

Examples

if (FALSE) {
# Authenticate first:
wt_auth()
a_camera_project <- wt_download_report(
project_id = 397, sensor_id = "CAM", reports = c("tag", "image_set"),
weather_cols = TRUE)

an_aru_project <- wt_download_report(
project_id = 47, sensor_id = "ARU", reports = c("main", "birdnet"),
weather_cols = TRUE)
}